주간 회고 조회

See merge request s11-final/S11P31E203!37

+++ server/src/main/java/com/e203/project/repository/ProjectRepository.java
@@ -1,8 +1,7 @@
 package com.e203.project.repository;
 
+import com.e203.project.entity.Project;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.e203.project.entity.Project;
-
-public interface ProjectRepository extends JpaRepository<Project, Long> {
+public interface ProjectRepository extends JpaRepository<Project, Integer> {
 }
